+++ Sources/editor/EditorTriggerAreaMode.cpp
@@ -79,7 +79,7 @@ namespace aga
                     out.push_back (yPoint);
                 }
 
-                al_draw_polygon (out.data (), it->second.Points.size (), 0, COLOR_DARKBLUE, 2, 0);
+                al_draw_polygon (out.data (), (int)it->second.Points.size (), 0, COLOR_DARKBLUE, 2, 0);
 
                 Point min{ std::numeric_limits<int>::max (), std::numeric_limits<int>::max () };
                 Point max{ std::numeric_limits<int>::min (), std::numeric_limits<int>::min () };
